refactor(api): extract dispatchChestTransfer to dedupe deposit and withdraw scaffolding
