+++ api/src/main/kotlin/dev/gvart/genesara/api/internal/mcp/tools/chest/WithdrawFromChestTool.kt
@@ -1,6 +1,5 @@
 package dev.gvart.genesara.api.internal.mcp.tools.chest
 
-import dev.gvart.genesara.api.internal.mcp.context.AgentContextHolder
 import dev.gvart.genesara.api.internal.mcp.presence.AgentActivityTracker
 import dev.gvart.genesara.api.internal.mcp.presence.touchActivity
 import dev.gvart.genesara.engine.TickClock
@@ -11,7 +10,6 @@ import org.springframework.ai.chat.model.ToolContext
 import org.springframework.ai.tool.annotation.Tool
 import org.springframework.ai.tool.annotation.ToolParam
 import org.springframework.stereotype.Component
-import java.util.UUID
 
 @Component
 internal class WithdrawFromChestTool(
@@ -36,28 +34,13 @@ internal class WithdrawFromChestTool(
         toolContext: ToolContext,
     ): ChestTransferResponse {
         touchActivity(toolContext, activity, "withdraw_from_chest")
-        val chestUuid = runCatching { UUID.fromString(chestId) }.getOrNull()
-            ?: return ChestTransferResponse.rejected(
-                chestId = chestId,
-                itemId = itemId,
+        return dispatchChestTransfer(chestId, itemId, quantity, world, engine) { chestUuid, agent ->
+            WorldCommand.WithdrawFromChest(
+                agent = agent,
+                chestId = chestUuid,
+                item = ItemId(itemId),
                 quantity = quantity,
-                reason = "bad_chest_id",
-                detail = "chestId must be a UUID",
             )
-        val agent = AgentContextHolder.current()
-        val command = WorldCommand.WithdrawFromChest(
-            agent = agent,
-            chestId = chestUuid,
-            item = ItemId(itemId),
-            quantity = quantity,
-        )
-        val appliesAtTick = world.submit(command, appliesAtTick = engine.currentTick() + 1)
-        return ChestTransferResponse.queued(
-            commandId = command.commandId,
-            appliesAtTick = appliesAtTick,
-            chestId = chestUuid,
-            itemId = itemId,
-            quantity = quantity,
-        )
+        }
     }
 }
